requirements.txt
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
ffmpeg-python==0.2.0
ffprobe==0.5
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-marshmallow==0.14.0
Flask-MySQLdb==0.2.0
flask-restplus==0.13.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
mysqlclient==2.0.3
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.18.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.23
typing-extensions==3.10.0.2
Werkzeug==0.16.1
zipp==3.5.0
